requirements.txt
streamlit
PyMuPDF
pandas
openpyxl
python-pptx
beautifulsoup4
faiss-cpu
sentence-transformers
google-generativeai
